requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pyyaml>=6.0.1

# PDF Processing
PyPDF2>=3.0.0

# LLM and Embeddings
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2

# Vector Store
faiss-cpu>=1.7.4

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Progress bars and CLI
tqdm>=4.65.0
click>=8.1.7

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Type checking and linting
mypy>=1.7.0
black>=23.11.0
isort>=5.12.0

# Utilities
requests>=2.31.0
typing-extensions>=4.8.0
